Python/mastopost/pyproject.toml
[project]
name = "mastopost"
version = "0.1.0"
description = "Add your description here"
readme = "README.md"
requires-python = ">=3.12"
dependencies = [
    "mastodon-py>=1.8.1",
    "python-dotenv>=1.0.1",
]
